django/pyproject.toml
[tool.poetry]
name = "podscription"
version = "0.1.0"
description = ""
authors = ["Zach Bellay <zachbellay@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.3"
psycopg2 = "^2.9.5"
whitenoise = "^6.2.0"
gunicorn = "^20.1.0"
django-ninja = "^0.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"